rename: convert project structure for porting
convert the structure of multiple files for porting from CRA to gatsby
(to use gatsby, the "page_components.js" files must be located directly under the "pages" folder.)
remove unneccessary image files and favicons

NOTE: The project in the current commit will NOT compile properly
because there is no additional modification of the source code to properly change the file structure.
This commit only records simple file moves and deletions.